+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 - SHACL diagram generation for the targetClasses (not shapes)
 - ability to visualize chosen classes as attributes (not as nodes)
 - support for cardinalities
+- support for hiding orphan nodes
 - support for filtering SHACL constraints  (e.g. required properties only)
 - export to PlantUML, TGF (yEd)